refactor: adopt mvc structure
